--- a/src/bounds/side.rs
+++ b/src/bounds/side.rs
@@ -3,36 +3,30 @@ use std::cmp::Ordering;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct Side {
-    // Pack positive and negative fields into a single u64:
-    // - bit 63: is_negative flag
-    // - bits 0-62: value (supports values up to 2^62)
-    packed: u64,
+    value: usize,
+    is_negative: bool,
 }
 
 impl std::fmt::Display for Side {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        if self.is_negative() {
-            write!(f, "-{}", self.value_unchecked() + 1)
+        if self.is_negative {
+            write!(f, "-{}", self.value + 1)
         } else {
-            write!(f, "{}", self.value_unchecked() + 1)
+            write!(f, "{}", self.value + 1)
         }
     }
 }
 
 impl Side {
-    const NEGATIVE_FLAG: u64 = 1u64 << 63;
-    const VALUE_MASK: u64 = !Self::NEGATIVE_FLAG;
-    const MAX_VALUE: usize = (Self::VALUE_MASK) as usize; // 2^63 - 1
-
     /**
      * Create a new Side, using the provided
      * value as-is. The value must be 0-indexed.
      */
     #[inline(always)]
-    pub fn with_pos_value(value: usize) -> Self {
-        debug_assert!(value <= (Self::MAX_VALUE), "Value too large");
+    pub fn with_pos_value(value0idx: usize) -> Self {
         Self {
-            packed: (value as u64),
+            value: value0idx,
+            is_negative: false,
         }
     }
 
@@ -41,10 +35,10 @@ impl Side {
      * value as-is. The value must be 0-indexed.
      */
     #[inline(always)]
-    pub fn with_neg_value(value: usize) -> Self {
-        debug_assert!(value <= (Self::MAX_VALUE), "Value too large");
+    pub fn with_neg_value(value0idx: usize) -> Self {
         Self {
-            packed: (value as u64) | Self::NEGATIVE_FLAG,
+            value: value0idx,
+            is_negative: true,
         }
     }
 
@@ -55,29 +49,30 @@ impl Side {
     #[inline(always)]
     pub fn with_pos_inf() -> Self {
         Self {
-            packed: Self::MAX_VALUE as u64,
+            value: Self::max_right(),
+            is_negative: false,
         }
     }
 
     #[inline(always)]
     pub fn value_unchecked(&self) -> usize {
-        (self.packed & Self::VALUE_MASK) as usize
+        self.value
     }
 
     #[inline(always)]
     #[must_use]
     pub fn value(&self) -> (bool, usize) {
-        (self.is_negative(), self.value_unchecked())
+        (self.is_negative, self.value)
     }
 
     #[inline(always)]
     pub const fn max_right() -> usize {
-        Self::MAX_VALUE
+        usize::MAX
     }
 
     #[inline(always)]
     pub fn is_negative(&self) -> bool {
-        self.packed & Self::NEGATIVE_FLAG != 0
+        self.is_negative
     }
 
     pub fn from_str_left_bound(s: &str) -> Result<Self, anyhow::Error> {
@@ -90,7 +85,10 @@ impl Side {
 
     fn from_str(s: &str, is_left_bound: bool) -> Result<Self, anyhow::Error> {
         Ok(match s {
-            "" => Side::with_pos_value(if is_left_bound { 0 } else { Self::max_right() }),
+            "" => Side {
+                value: if is_left_bound { 0 } else { Self::max_right() },
+                is_negative: false,
+            },
             _ => {
                 let v = s
                     .parse::<isize>()
@@ -101,9 +99,15 @@ impl Side {
                 }
 
                 if v > 0 {
-                    Side::with_pos_value(usize::try_from(v.abs()).unwrap() - 1)
+                    Side {
+                        value: usize::try_from(v.abs()).unwrap() - 1,
+                        is_negative: false,
+                    }
                 } else {
-                    Side::with_neg_value(usize::try_from(v.abs()).unwrap() - 1)
+                    Side {
+                        value: usize::try_from(v.abs()).unwrap() - 1,
+                        is_negative: true,
+                    }
                 }
             }
         })
@@ -113,15 +117,15 @@ impl Side {
 impl PartialOrd for Side {
     #[inline(always)]
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        if self.is_negative() ^ other.is_negative() {
+        if self.is_negative ^ other.is_negative {
             // We can't compare two sides with different sign
             return None;
         }
 
-        if self.is_negative() {
-            Some(other.value_unchecked().cmp(&self.value_unchecked()))
+        if self.is_negative {
+            Some(other.value.cmp(&self.value))
         } else {
-            Some(self.value_unchecked().cmp(&other.value_unchecked()))
+            Some(self.value.cmp(&other.value))
         }
     }
 }
@@ -133,18 +137,36 @@ mod tests {
     #[test]
     fn test_from_str_some() {
         assert_eq!(
-            Side::from_str_left_bound("42").unwrap().value(),
-            (false, 41)
+            Side::from_str_left_bound("42").unwrap(),
+            Side {
+                value: 41,
+                is_negative: false
+            }
         );
-        assert_eq!(Side::from_str_left_bound("-7").unwrap().value(), (true, 6));
+        assert_eq!(
+            Side::from_str_left_bound("-7").unwrap(),
+            Side {
+                value: 6,
+                is_negative: true
+            }
+        );
     }
 
     #[test]
     fn test_from_str_continue() {
-        assert_eq!(Side::from_str_left_bound("").unwrap().value(), (false, 0));
         assert_eq!(
-            Side::from_str_right_bound("").unwrap().value(),
-            (false, Side::max_right())
+            Side::from_str_left_bound("").unwrap(),
+            Side {
+                value: 0,
+                is_negative: false
+            }
+        );
+        assert_eq!(
+            Side::from_str_right_bound("").unwrap(),
+            Side {
+                value: usize::MAX,
+                is_negative: false
+            }
         );
     }