Side is now accessed through implemented methods, which offers more wiggle room for future refactoring (maybe generics?).
Side contains zero-indexed values, so we can skip some arithmetic here and there (and de facto way more indexes, from i32 to usize, but that's hardly going to be used, maybe reading binaries...).


Despite the big changes the difference in performance (wall clock time) is minimal.
Overall, fewer instructions, fewer L1 cache miss, minimally better branch prediction


```=== Performance Comparison ===

=== tuc-master ===

 Performance counter stats for 'taskset -c 1 ./tuc-master -d    -f 2,-2 /tmp/data-multichar.txt' (10 runs):

     5,518,124,703      cycles:u                                                                ( +-  2.98% )  (39.94%)
    12,605,719,278      instructions:u                   #    2.28  insn per cycle
                                                  #    0.01  stalled cycles per insn     ( +-  0.05% )  (39.93%)
     4,421,387,714      L1-dcache-loads:u                                                       ( +-  0.20% )  (39.94%)
        11,564,142      L1-dcache-load-misses:u          #    0.26% of all L1-dcache accesses   ( +-  3.36% )  (39.96%)
         3,276,799      L1-icache-load-misses:u                                                 ( +-  3.29% )  (40.00%)
         2,940,505      dTLB-loads:u                                                            ( +-  0.12% )  (40.03%)
         2,749,414      dTLB-load-misses:u               #   93.50% of all dTLB cache accesses  ( +-  0.09% )  (40.04%)
             4,740      iTLB-loads:u                                                            ( +-  3.71% )  (40.05%)
             6,468      iTLB-load-misses:u               #  136.46% of all iTLB cache accesses  ( +-  4.23% )  (40.04%)
     2,605,332,603      branches:u                                                              ( +-  0.05% )  (40.04%)
        10,553,217      branch-misses:u                  #    0.41% of all branches             ( +-  0.78% )  (40.04%)
        97,896,460      stalled-cycles-frontend:u        #    1.77% frontend cycles idle        ( +- 14.86% )  (40.04%)
       105,549,553      stalled-cycles-backend:u         #    1.91% backend cycles idle         ( +- 11.01% )  (40.02%)
        23,464,926      cache-misses:u                   #   10.62% of all cache refs           ( +-  1.67% )  (39.98%)
       220,855,256      cache-references:u                                                      ( +-  0.78% )  (39.95%)

            2.8589 +- 0.0406 seconds time elapsed  ( +-  1.42% )

----------------------------------------

=== tuc-10933d3 ===

 Performance counter stats for 'taskset -c 1 ./tuc-10933d3 -d    -f 2,-2 /tmp/data-multichar.txt' (10 runs):

     5,389,054,568      cycles:u                                                                ( +-  0.19% )  (40.01%)
    12,502,447,571      instructions:u                   #    2.32  insn per cycle
                                                  #    0.01  stalled cycles per insn     ( +-  0.04% )  (40.01%)
     4,437,491,477      L1-dcache-loads:u                                                       ( +-  0.25% )  (40.01%)
        15,938,639      L1-dcache-load-misses:u          #    0.36% of all L1-dcache accesses   ( +- 27.76% )  (40.01%)
           903,479      L1-icache-load-misses:u                                                 ( +-  0.73% )  (40.00%)
         2,950,686      dTLB-loads:u                                                            ( +-  0.17% )  (40.01%)
         2,749,183      dTLB-load-misses:u               #   93.17% of all dTLB cache accesses  ( +-  0.08% )  (39.99%)
             6,003      iTLB-loads:u                                                            ( +-  3.81% )  (39.99%)
             5,822      iTLB-load-misses:u               #   96.98% of all iTLB cache accesses  ( +-  2.58% )  (39.99%)
     2,564,811,188      branches:u                                                              ( +-  0.03% )  (40.00%)
        10,267,471      branch-misses:u                  #    0.40% of all branches             ( +-  0.72% )  (40.00%)
        83,804,303      stalled-cycles-frontend:u        #    1.56% frontend cycles idle        ( +-  4.25% )  (39.99%)
       118,104,976      stalled-cycles-backend:u         #    2.19% backend cycles idle         ( +-  0.36% )  (40.00%)
        23,388,253      cache-misses:u                   #   10.57% of all cache refs           ( +-  1.46% )  (39.99%)
       221,263,948      cache-references:u                                                      ( +-  3.11% )  (39.99%)

           2.81932 +- 0.00194 seconds time elapsed  ( +-  0.07% )

----------------------------------------

=== Alternative: Focused Measurements ===
--- Core Execution ---
tuc-master:

 Performance counter stats for 'taskset -c 1 ./tuc-master -d    -f 2,-2 /tmp/data-multichar.txt' (10 runs):

     5,356,010,384      cycles:u                                                                ( +-  0.19% )
    12,614,350,891      instructions:u                   #    2.36  insn per cycle              ( +-  0.00% )
     2,604,224,490      branches:u                                                              ( +-  0.00% )
        10,829,364      branch-misses:u                  #    0.42% of all branches             ( +-  1.22% )

           2.80035 +- 0.00147 seconds time elapsed  ( +-  0.05% )


tuc-10933d3:

 Performance counter stats for 'taskset -c 1 ./tuc-10933d3 -d    -f 2,-2 /tmp/data-multichar.txt' (10 runs):

     5,373,427,610      cycles:u                                                                ( +-  0.30% )
    12,516,348,136      instructions:u                   #    2.33  insn per cycle              ( +-  0.00% )
     2,562,223,892      branches:u                                                              ( +-  0.00% )
        10,530,362      branch-misses:u                  #    0.41% of all branches             ( +-  1.10% )

           2.79172 +- 0.00371 seconds time elapsed  ( +-  0.13% )


--- Memory Hierarchy ---
tuc-master:

 Performance counter stats for 'taskset -c 1 ./tuc-master -d    -f 2,-2 /tmp/data-multichar.txt' (10 runs):

     4,428,202,965      L1-dcache-loads:u                                                       ( +-  0.05% )
        12,467,235      L1-dcache-load-misses:u          #    0.28% of all L1-dcache accesses   ( +- 10.38% )
       222,669,633      cache-references:u                                                      ( +-  0.75% )
        23,013,473      cache-misses:u                   #   10.34% of all cache refs           ( +-  0.40% )

          2.801282 +- 0.000530 seconds time elapsed  ( +-  0.02% )


tuc-10933d3:

 Performance counter stats for 'taskset -c 1 ./tuc-10933d3 -d    -f 2,-2 /tmp/data-multichar.txt' (10 runs):

     4,447,668,339      L1-dcache-loads:u                                                       ( +-  0.17% )
        10,393,877      L1-dcache-load-misses:u          #    0.23% of all L1-dcache accesses   ( +-  3.68% )
       214,839,870      cache-references:u                                                      ( +-  0.29% )
        23,072,819      cache-misses:u                   #   10.74% of all cache refs           ( +-  0.90% )

           2.80134 +- 0.00146 seconds time elapsed  ( +-  0.05% )


--- Pipeline Efficiency ---
tuc-master:

 Performance counter stats for 'taskset -c 1 ./tuc-master -d    -f 2,-2 /tmp/data-multichar.txt' (10 runs):

       101,718,499      stalled-cycles-frontend:u        #    1.91% frontend cycles idle        ( +-  6.74% )
       113,477,886      stalled-cycles-backend:u         #    2.13% backend cycles idle         ( +-  3.16% )
     5,333,206,750      cycles:u                                                                ( +-  0.26% )
    12,614,350,610      instructions:u                   #    2.37  insn per cycle
                                                  #    0.01  stalled cycles per insn     ( +-  0.00% )

           2.80065 +- 0.00150 seconds time elapsed  ( +-  0.05% )


tuc-10933d3:

 Performance counter stats for 'taskset -c 1 ./tuc-10933d3 -d    -f 2,-2 /tmp/data-multichar.txt' (10 runs):

        81,914,003      stalled-cycles-frontend:u        #    1.53% frontend cycles idle        ( +-  1.64% )
       118,855,012      stalled-cycles-backend:u         #    2.22% backend cycles idle         ( +-  0.11% )
     5,346,344,656      cycles:u                                                                ( +-  0.05% )
    12,516,348,264      instructions:u                   #    2.34  insn per cycle
                                                  #    0.01  stalled cycles per insn     ( +-  0.00% )

          2.803538 +- 0.000418 seconds time elapsed  ( +-  0.01% )

```